requirements.txt
fastapi
fastmcp
jaydebeapi
JPype1
python-dotenv
pydantic
uvicorn
mcp
